Do not load autoinstrumentation assembly on splashscreen (#1253)
