+++ test/test-applications/integrations/TestApplication.AspNet/Controllers/HomeController.cs
@@ -21,6 +21,7 @@ using System.Linq;
 using System.Reflection;
 using System.Web.Mvc;
 using TestApplication.AspNet.Helpers;
+using TestApplication.Shared;
 
 namespace TestApplication.AspNet.Controllers;
 
@@ -28,25 +29,10 @@ public class HomeController : Controller
 {
     public ActionResult Index()
     {
-        var prefixes = new[] { "COR_", "CORECLR_", "DOTNET_", "OTEL_" };
-
-        var envVars = from envVar in Environment.GetEnvironmentVariables().Cast<DictionaryEntry>()
-                      from prefix in prefixes
-                      let key = (envVar.Key as string)?.ToUpperInvariant()
-                      let value = envVar.Value as string
-                      where key.StartsWith(prefix)
-                      orderby key
-                      select new KeyValuePair<string, string>(key, value);
-
-        var instrumentationType =
-            Type.GetType("OpenTelemetry.AutoInstrumentation.Instrumentation, OpenTelemetry.AutoInstrumentation");
+        var envVars = ProfilerHelper.GetEnvironmentConfiguration();
 
         ViewBag.EnvVars = envVars;
         ViewBag.HasEnvVars = envVars.Any();
-        ViewBag.ProfilerAttached =
-            instrumentationType?.GetProperty("ProfilerAttached", BindingFlags.Public | BindingFlags.Static)
-                ?.GetValue(null) ?? false;
-        ViewBag.TracerAssemblyLocation = instrumentationType?.Assembly.Location;
         ViewBag.TracerAssemblies = AssembliesHelper.GetLoadedTracesAssemblies();
         ViewBag.AllAssemblies = AssembliesHelper.GetLoadedAssemblies();
